+++ packages/alfa-dom/jsx-runtime.ts
@@ -1,3 +1,1 @@
-import { jsx } from "./src/jsx";
-export { jsx };
-export import JSX = jsx.JSX;
+export * from "./src/jsx-runtime";