+++ packages/alfa-rules/test/sia-r16/rule.spec.tsx
@@ -96,7 +96,7 @@ test(`evaluate() fails a focusable <div> element with a role of separator and no
 });
 
 test("evaluate() is inapplicable to elements that are not exposed", async (t) => {
-  const target = <div role="combobox" style={{ display: "none" }}></div>;
+  const target = <div role="combobox" style={{ display: "none" }} />;
 
   const document = Document.of([target]);