+++ packages/alfa-style/test/property/inset-block.spec.tsx
@@ -8,7 +8,7 @@ import { Style } from "../../src/style";
 const device = Device.standard();
 
 test("#cascaded() parses `inset-block: 1px 2px", (t) => {
-  const element = <div style={{ insetBlock: "1px 2px" }}></div>;
+  const element = <div style={{ insetBlock: "1px 2px" }} />;
 
   const style = Style.from(element, device);
 
@@ -32,7 +32,7 @@ test("#cascaded() parses `inset-block: 1px 2px", (t) => {
 });
 
 test("#cascaded() parses `inset-block: 1px", (t) => {
-  const element = <div style={{ insetBlock: "1px" }}></div>;
+  const element = <div style={{ insetBlock: "1px" }} />;
 
   const style = Style.from(element, device);