+++ packages/alfa-style/test/property/text-indent.spec.tsx
@@ -8,7 +8,7 @@ import { Style } from "../../src/style";
 const device = Device.standard();
 
 test("#cascaded() parses `text-indent: 5px`", (t) => {
-  const element = <div style={{ textIndent: "5px" }}></div>;
+  const element = <div style={{ textIndent: "5px" }} />;
 
   const style = Style.from(element, device);